+++ Repository/IGenericRepository.cs
@@ -5,7 +5,7 @@ namespace Dating.API.Repository
 {
     public interface IGenericRepository<TEntity> where TEntity: class
     {
-        Task<TEntity> Get(int Id);
+        Task<TEntity> Get(long Id);
         Task<IEnumerable<TEntity>> GetAll();
         void Add(TEntity entity);
         void Delete(TEntity entity);
